fix: Changed type of EquipmentsProps value

+++ enkanetwork/model/assets.py
@@ -114,7 +114,7 @@ class AritfactProps(BaseModel):
     id: int = 0
     type: str = Field('', alias='propType')
     digit: str = Field('DIGIT', alias='propDigit')
-    value: int = Field(0, alias='propValue')
+    value: float = Field(0, alias='propValue')
 
 
 class CharacterAsset(BaseModel):
